Do not clear values
This fixes #12
If a value was previously entered and removed from the displayed view, the value is not cleared.
The value is still displayed (in italic) but is not selectable on the dropdown.

It prevents the control to wipe-out the value and acts like the out of the box feature.